+++ examples/example-3/larger_bow_tie_array_driver.m
@@ -45,10 +45,10 @@ Const.SUNEMcbfmstrfilename     = 'cbfm_larger_bow_tie_array.str';
 % TO-DO: Setup some documentation for this - also assign default values if they are not
 % defined explicitely here.
 Const.no_mutual_coupling_array = false; % Deactivate coupling between domains.
-Const.calcSecMBFs = true;     % For MBF based solvers
+Const.calcSecMBFs = false;      % For MBF based solvers
 Const.useMBFreduction = true;  % SVD applied after the MBFs are generated to retain an orthonormal set
-Const.MBFthreshold = 10;     % Threshold used for the SVD reduction of the MBFs
-Const.IFBalg = 7;              % Jacobi iterations (7). Adaptive MBF (14).
+Const.MBFthreshold = 10;       % Threshold used for the SVD reduction of the MBFs
+Const.IFBalg = 14;             % Jacobi iterations (7). Adaptive MBF (14).
 Const.IFB_iterations = 10;     % Number of Jacobi iterations. (TO-DO: Ellaborate special meaning, e.g. -1)
                                % which then looks at Const.IFB_convergence_threshold_percentage;
 Const.IFB_convergence_threshold_percentage = 1E-3;                                
@@ -57,7 +57,6 @@ Const.IFB_CBFs = -1;           % TO-DO: Recheck this - essentially for the Adapt
 Const.IFB_debug = 1;
 Const.cache_Z0_V0 = false;     % Precompute the Z0 and V0 terms
 Const.use_DGFM_start = false;  % Use the DGFM to calculate the initial (0th) solution
-
 
 % --------------------------------------------------------------------------------------------------
 % Read the MoM matrix equation from the file